Added code to notebook_to_pdf to pass in all notebook metadata to nbconvert
Not sure why, but nbconvert's from_notebook_node function seems to
ignore all of the metadata in the notebook node that is passed in. So,
I added a bit of code to pull in all of the metadata and pass it to
the conversion function. This allows the user to set things like the
list of authors for a document. It also allows the user to override
the notebook title as the document title. This gives the user to the
ability to have file names such as this_is_my_document.ipynb, while
still getting clean titles like "This is My Document" via the notebook
metadata.